requirements.txt
flask
flask_cors
cryptography
python-dotenv
pytz
pymongo
werkzeug
flask-socketio